// client/my-app/src/constant.ts
export enum RoutePath {                             ////// определение всех наших маршрутов
    HOME = "/",
    FORGOT_PASSWORD = "/forgot_password",
    LOG_IN = "/login",
    LOG_OUT = "/log_out",
    NOTIFICATIONS = "/notifications",
    SETTINGS = "/settings",
    MESSAGES="/messages",
    REMOVED_MESSAGES="/removed_messages",
    CREATE_MESSAGE="/create_message",
    REGISTRATION="/registration"
}